refactor: Rename /geometries to /Geometry
